zephyr: Update GOOGLE_RTC_AUDIO_PROCESSING
Unbreak the Zephyr build when this is enabled, and add the needed bits
to produce a working executable.

This is mostly just a recapitulation of the existing integration,
which means that it's manually pulling in bits from the Cadence
toolchain it needs.  Really this is nothing more than a C++ library
and it shouldn't have to do anything other than enable CONFIG_CPP=y in
the build.  But the upstream Zephyr C++ toolchain integration for
xcc/xt-clang isn't present yet.

Signed-off-by: Andy Ross <andyross@google.com>